Create groupBirthdays
In a large group of people it is probable that many might have the same birthday. This program tests that by asking the user for a number of people, generates birthdays for them and then determines whether or not there are matches. It runs the simulation 100,000 times to create a probability for the group size.